requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.9
billiard==3.6.4.0
black==22.6.0
celery==5.2.7
certifi==2022.9.24
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
dill==0.3.5.1
distlib==0.3.6
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.0
isort==5.10.1
kombu==5.2.4
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
pipenv==2022.10.12
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2==2.9.4
PyJWT==2.6.0
pylint==2.15.0
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dotenv==0.21.0
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.4
tzdata==2022.5
vine==5.0.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
wcwidth==0.2.5
wrapt==1.14.1
